requirements.txt
Django==4.2.13
djangorestframework==3.14.0
python-decouple==3.8
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
celery==5.3.6
redis==5.0.1
requests==2.31.0
django-filter==24.1
python-dateutil==2.8.2